requirement.txt
# requirements.txt
httpx>=0.25.0
asyncio
requests
h2
selenium
chromedriver-autoinstaller
pyOpenSSL
dnspython
tqdm
colorama
aioquic
